Implemented systemtools.TestCase.
