+++ abjad/tools/test/test_abjad___init__.py
@@ -25,6 +25,7 @@ ignored_classes = (
     datastructuretools.Enumeration,
     systemtools.StorageFormatAgent,
     systemtools.FormatSpecification,
+    systemtools.TestCase,
     )
 
 classes = documentationtools.list_all_abjad_classes(
